Fixed ListOptionsContainer.tsx problem. Now options container scrolls whenever screen is too tiny.
