+++ src/assets/styles/App.css
@@ -27,6 +27,9 @@ body {
 
 @media screen and (max-width: 450px) {
     body {
+        overflow: hidden;
+        width: 100%;
+        height: 100%;
         background-color: var(--secondary-background-color);
     }
 }
