+++ src/assets/styles/variables.css
@@ -17,7 +17,8 @@
     --desktop-content-container-width: 1300px;
     --tablet-content-container-width: 1000px;
     --mobile-content-container-width: 450px;
-    --content-padding: 40px;
+    --big-screen-content-padding: 40px;
+    --mobile-screen-content-padding: 20px;
 }
 
 /*--------FONTS----------------------------------------------------*/
@@ -30,4 +31,14 @@
     --main-text-bigger-bolder: 800 16px/1.2 'Be Vietnam Pro', sans-serif;
     --subscript-text-bold: 800 10px/1 'Be Vietnam Pro', sans-serif;
     --subscript-text: 400 10px/1 'Be Vietnam Pro', sans-serif;
+
+    --heading-text-mobile: 800 28px/1.4 'Be Vietnam Pro', sans-serif;
+    --subheading-text-bigger-mobile: 800 20px/1.4 'Be Vietnam Pro', sans-serif;
+    --subheading-text-mobile: 800 17px/1.4 'Be Vietnam Pro', sans-serif;
+    --main-text-mobile: 400 12px/1.2 'Be Vietnam Pro', sans-serif;
+    --main-text-bold-mobile: 600 12px/1.2 'Be Vietnam Pro', sans-serif;
+    --main-text-bolder-mobile: 800 12px/1.2 'Be Vietnam Pro', sans-serif;
+    --main-text-bigger-bolder-mobile: 800 13px/1.2 'Be Vietnam Pro', sans-serif;
+    --subscript-text-bold-mobile: 800 8px/1 'Be Vietnam Pro', sans-serif;
+    --subscript-text-mobile: 400 8px/1 'Be Vietnam Pro', sans-serif;
 }
