+++ src/components/common/AppCategoryOption/AppCategoryOption.tsx
@@ -1,12 +1,12 @@
-import './CategoryOptionDataContainer.css'
+import './AppCategoryOption.css'
 import { useCategoryOption } from "../../../states/categoryListOptionState.ts";
 
-export const CategoryOptionDataContainer = () => {
+export const AppCategoryOption = () => {
     const categoryOptions = useCategoryOption((state) => state.categoryOptions)
     const selectedOptionIndex = categoryOptions.findIndex((option) => option.isSelected)
 
     return (
-        <div>
+        <div className="app-category-option__container">
             {categoryOptions[selectedOptionIndex].label}
         </div>
     )
