+++ src/components/features/ListOptionFeature/ListOptionsContainer/ListOptionsContainer.css
@@ -1,0 +1,38 @@
+.category-options__container {
+    width: 100%;
+    height: fit-content;
+}
+
+hr.solid {
+    border-top: 1px solid var(--hoovered-light-gray-color);
+    padding-bottom: 15px;
+}
+
+.category-options__label {
+    width: 100%;
+    font: var(--main-text-bigger-bolder);
+    padding-bottom: 5px;
+}
+
+@media screen and (max-width: 450px) {
+    .category-options__label {
+        font: var(--main-text-bolder-mobile);
+    }
+
+    .category-options__list-container {
+        position: absolute;
+        width: 100%;
+        overflow-x: scroll;
+        margin-left: -20px;
+        padding-left: 20px;
+    }
+
+    .category-options__list-container::-webkit-scrollbar {
+        display: none;
+    }
+}
+
+.category-options__list-container {
+    width: 100%;
+    height: fit-content;
+}
